Moving HTTP code out of scan_engine and moving mkschema into the package
