+++ tests/core.py
@@ -13,8 +13,10 @@ cli_args = {'category': None,
             'dryrun': False,
             'fingerprint': False,
             'log': None,
+            'mkcred': False,
             'name': None,
             'nmap': None,
+            'protocols': True,
             'proxy': None,
             'output': None,
             'subnet': None,
